% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.local.minimum.R
\name{first.local.minimum}
\alias{first.local.minimum}
\title{Index of the first local minimum.}
\usage{
first.local.minimum(x)
}
\arguments{
\item{x}{vector.}
}
\value{
the index of the first local minimum of \code{x}.
}
\description{
This function computes the index of the first local minimum.
}
\examples{

v<-rnorm(30)
out<-first.local.minimum(v)

}
\references{
Kraemer, N., Sugiyama M. (2011). "The Degrees of Freedom of
Partial Least Squares Regression". Journal of the American Statistical
Association. ahead of print 106 (494)
\url{https://www.tandfonline.com/doi/abs/10.1198/jasa.2011.tm10107}
}
\author{
Nicole Kraemer
}
\keyword{math}
