% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization of vectors}
\usage{
normalize(v, w = NULL)
}
\arguments{
\item{v}{vector}

\item{w}{optional vector}
}
\value{
\item{v}{normalized \code{v}} \item{w}{normalized \code{w}}
}
\description{
Normalization of vectors.
}
\details{
The vector \code{v} is normalized to length 1. If \code{w} is given, it is
normalized by the length of \code{v}.
}
\examples{

v<-rnorm(5)
w<-rnorm(10)
dummy<-normalize(v,w)

}
\author{
Nicole Kraemer, Mikio L. Braun
}
\keyword{math}
