% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr.r
\name{tr}
\alias{tr}
\title{Trace of a matrix}
\usage{
tr(M)
}
\arguments{
\item{M}{square matrix}
}
\value{
The trace of the matrix M.
}
\description{
This function computes the trace of a matrix.
}
\examples{

M<-matrix(rnorm(8*8),ncol=8)
tr.M<-tr(M)

}
\author{
Nicole Kraemer
}
\keyword{math}
