% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krylov.r
\name{krylov}
\alias{krylov}
\title{Krylov sequence}
\usage{
krylov(A, b, m)
}
\arguments{
\item{A}{square matrix of dimension p x p.}

\item{b}{vector of length p}

\item{m}{length of the Krylov sequence}
}
\value{
A matrix of size p x m containing the sequence b,Ab,..., A^(m-1)b.
}
\description{
This function computes the Krylov sequence of a matrix and a vector.
}
\examples{

A<-matrix(rnorm(8*8),ncol=8)
b<-rnorm(8)
K<-krylov(A,b,4)

}
\author{
Nicole Kraemer
}
\keyword{math}
