\name{plsdof-package}
\Rdversion{1.1}
\alias{plsdof-package}
\alias{plsdof}
\docType{package}
\title{
Degrees of Freedom and Confidence Intervals for Partial Least Squares Regression
}
\description{The \code{plsdof} provides Degrees of Freedom estimates and confidence intervals for Partial Least Squares Regression. The package also allows model selection based on various information criteria (aic, bic, gmdl) and based on cross-validation.
}
\details{
\tabular{ll}{
Package: \tab plsdof\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2010-02-22\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{Nicole Kraemer, Mikio L. Braun

Maintainer: Nicole Kraemer <nkraemer@cs.tu-berlin.de>
}
\references{

Kraemer, N., Sugiyama M. (2010). "The Degrees of Freedom of Partial Least Squares Regression". preprint, \url{http://arxiv.org/abs/1002.4112}


Kraemer, N., Braun, M.L. (2007) "Kernelizing PLS, Degrees of Freedom, and Efficient Model Selection", Proceedings of the 24th International Conference on Machine Learning, Omni Press, 441 - 448 

}


\seealso{
\code{\link{pls.model}}, \code{\link{pls.cv}}, \code{\link{pls.ic}}

}
\keyword{ package }

\examples{
# to do
}
