\name{plsdof-package}
\Rdversion{1.1}
\alias{plsdof-package}
\alias{plsdof}
\docType{package}
\title{
Degrees of Freedom of Partial Least Squares Regression
}
\description{The \code{psldof} package provides Degrees of Freedom estimates for Kernel Partial Least Squares Regression and its kernel extension. The package also provides model selection based on various information criteria (aic, bic, gmdl) and based on cross-validation.
}
\details{
\tabular{ll}{
Package: \tab plsdof\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-11-16\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{Nicole Kraemer, Mikio L. Braun

Maintainer: Nicole Kraemer <nkraemer@cs.tu-berlin.de>
}
\references{

Kraemer, N., Braun, M.L. (2007) "Kernelizing PLS, Degrees of Freedom, and Efficient Model Selection", Proceedings of the 24th International Conference on Machine Learning, Omni Press, 441 - 448 

}


\seealso{
\code{\link{kernel.pls}}, \code{\link{kernel.pls.cv}}, \code{\link{kernel.pls.ic}}

}
\keyword{ package }

\examples{
# to do
}
